test: Fixup implicit double casts in test code
We use the clang options -Wdouble-promotion -Werror=double-promotion
to discover places where picolibc is unintentionally doing
computations using doubles instead of floats. With that, any
intentional casts must be explicit, so this adds those to the test
code.

+++ test/fenv.c
@@ -116,7 +116,7 @@ static int
 report(char *expr, test_t v, int e, int exception, int oexception)
 {
 	printf("%-20.20s: ", expr);
-	printf("%8g ", v);
+	printf("%8g ", (double) v);
 	printf("(e expect %s", e_to_str(exception));
 	if (oexception)
 		printf(" or %s", e_to_str(oexception));
